Rearrange Code and Add Release Build Mode
The release build mode uses a generic key, it's called generickey or something and the password for the keystore is password and the password for the key itself is password123. I don't know if I should be doing this, but at least it got Android Studio to shut up about there not being a signature on the release.